Set locale ENV, update startup script.

+++ root/etc/my_init.d/startup.sh
@@ -2,10 +2,11 @@
 
 HOME=~project
 
-# Copy default sshd config
-if [ ! -h /etc/ssh/sshd_config  ]; then
-    # Regenerate ssh keys
+# Generate SSH host keys for first time use
+# And also generate password for the default user
+if [ ! -f /sshkey_generated ]; then
     /etc/my_init.d/00_regen_ssh_host_keys.sh > /dev/nul 2>&1
+    touch /sshkey_generated
 
     # Set the user passwd
     USER_PASSWORD=`pwgen -c -n -1 10`
@@ -13,7 +14,10 @@ if [ ! -h /etc/ssh/sshd_config  ]; then
     echo -e "${USER_PASSWORD}\n${USER_PASSWORD}" | passwd project > /dev/nul 2>&1
     echo "Password for project user: ${USER_PASSWORD}"
     echo "Please change your password!"
+fi
 
+# Copy default sshd config
+if [ ! -h /etc/ssh/sshd_config ]; then
     # Create symbolic link for sshd_config
     rm -rf /etc/ssh/sshd_config
     cp /etc/ssh/sshd_config.original ${HOME}/sshd_config
